refactor: DTO 패키지 병합
Response, Request 패키지 DTO 패키지로 병합.

Related to: #8
