+++ src/main/java/juwoncode/commonblogproject/dto/MemberResponse.java
@@ -1,0 +1,5 @@
+package juwoncode.commonblogproject.dto;
+
+public class MemberResponse {
+
+}
