--- src/main/java/juwoncode/commonblogproject/response/MemberResponse.java
@@ -1,5 +1,0 @@
-package juwoncode.commonblogproject.response;
-
-public class MemberResponse {
-
-}
